Refactor TaskContext, useReducer
