+++ src/components/tasks/TaskItem.jsx
@@ -1,12 +1,91 @@
+import { useContext } from "react";
 import { FaTrash, FaEdit, FaCalendarCheck, FaCalendarTimes } from "react-icons/fa";
-import { completeTask, editTask, removeTask } from "../../actions/taskActions";
+import TaskContext from "../../context/TaskContext";
 
 export default function TaskItem(props) {
+  const { tasks, dispatch } = useContext(TaskContext);
+
+  const completeTask = (task) => {
+    let toggled = {};
+    tasks.map((item) => {
+      if (item.id === task.id) {
+        toggled = {
+          ...item,
+          status: !task.status,
+        };
+      }
+    });
+    dispatch({
+      type: "TASK_UPDATE",
+      payload: toggled,
+    });
+  };
+
+  const editTask = (task) => {
+    // Open form on mobile devices
+    const formMobile = document.querySelector('.my-form-mobile');
+    formMobile.closest('.modal').classList.add('is-active');
+    document.querySelector('html').classList.add('is-clipped');
+
+    dispatch({
+      type: "FORM_EDIT",
+      payload: task,
+    });
+
+    const editTitle = task.title;
+    document.getElementById('title').value = editTitle;
+    document.getElementById('title-mobile').value = editTitle;
+    dispatch({
+      type: "FORM_SET_TITLE",
+      payload: editTitle,
+    });
+
+    const editDesciption = task.description;
+    document.getElementById('description').value = editDesciption;
+    document.getElementById('description-mobile').value = editDesciption;
+    dispatch({
+      type: "FORM_SET_DESCRIPTION",
+      payload: editDesciption,
+    });
+
+    const editDatetime = new Date(task.datetime);
+    dispatch({
+      type: "FORM_SET_DATETIME",
+      payload: editDatetime,
+    });
+
+    const priorityField = document.getElementById('priority');
+    const priorityFieldMobile = document.getElementById('priority-mobile');
+    if (task.priority === 'high') {
+      priorityField.value = 3;
+      priorityFieldMobile.value = 3;
+    } else if (task.priority === 'medium') {
+      priorityField.value = 2;
+      priorityFieldMobile.value = 2;
+    } else {
+      priorityField.value = 1;
+      priorityFieldMobile.value = 1;
+    }
+    dispatch({
+      type: "FORM_SET_PRIORITY",
+      payload: priorityField.value,
+    });
+  };
+
+  const removeTask = (task) => {
+    if (confirm(`Remove ${task.title} from list?`)) {
+      dispatch({
+        type: "TASK_REMOVE",
+        payload: task,
+      })
+    }
+  };
+
   return (
     <div className="column is-one-third" style={props.style}>
       <div className="card has-background-light">
         <header className="card-header">
-          <p className="card-header-title">
+          <p className="card-header-title" style={props.task.status === true ? {textDecoration: "line-through"} : {}}>
             {props.task.title}
           </p>
           <button className="card-header-icon py-1 px-3" aria-label="complete task" title="Complete" onClick={() => completeTask(props.task)}>
